pyproject.toml
[project]
name = "yolo"
version = "0.2.0"
description = "Official Implementation of YOLO"
authors = [
  {name = "Shreyas Kamath", email = "shreyaskamathkm@gmail.com"}
]
maintainers = [
  {name = "Shreyas Kamath", email = "shreyaskamathkm@gmail.com"}
]

readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["yolo", "object detection", "computer vision"]
dependencies = [
    "einops",
    "faster-coco-eval",
    "graphviz",
    "hydra-core",
    "lightning",
    "loguru",
    "numpy",
    "opencv-python",
    "Pillow",
    "pycocotools",
    "requests",
    "rich",
    "torch",
    "torchvision",
    "wandb",
]

[project.optional-dependencies]
dev = [
    "yolo[docs]",
    "bump-my-version",
    "gradio",
    "pytest",
    "pytest-cov",
    "pre-commit",
    "tensorboard",
]
docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
    "pymdown-extensions",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["yolo*"]

[tool.setuptools]
package-data = {"yolo" = ["**/*.yaml"]}
include-package-data = true


[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools",
]

[project.scripts]
yolo = "yolo.cli:main"

[tool.bumpversion]
current_version = "0.2.0"
commit = true
tag = false
message = "chore: bump version {current_version} -> {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[project.urls]
Homepage = "https://github.com/shreyaskamathkm/yolo"
"Original Work" = "https://github.com/MultimediaTechLab/YOLO"
